% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R
\name{add_package}
\alias{add_package}
\title{Add a package to the packages.R file}
\usage{
add_package(package_name)
}
\arguments{
\item{package_name}{The name of the package to add.}
}
\description{
This function adds a package to the \code{packages.R} file in the user's project
root. This makes the package available for use in the targets pipeline.
}
\examples{
\dontrun{
add_package("ggplot2")
}
}
