% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tflow.R
\name{qmd_target}
\alias{qmd_target}
\title{qmd_target}
\usage{
qmd_target(target_name)
}
\arguments{
\item{target_name}{of a target to generate qmd target for.}
}
\value{
target text to the console.
}
\description{
Generate a target for a Quarto file
}
\author{
Robert M Flight
}
