% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dflow.R
\name{rmd_target}
\alias{rmd_target}
\title{use_rmd_target}
\usage{
rmd_target(file_path)
}
\arguments{
\item{file_path}{a file to generate target text for.}
}
\value{
target text to the console.
}
\description{
Generate a target for an R markdown file
}
\author{
Miles McBain
}
