% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dflow.R
\name{use_dflow}
\alias{use_dflow}
\title{use_dflow}
\usage{
use_dflow()
}
\value{
Nothing. Modifies your workspace.
}
\description{
Setup a dflow project
}
\details{
Creates files and directories according to the dflow template.
}
