% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dflow.R
\name{use_gitignore}
\alias{use_gitignore}
\title{use_gitignore}
\usage{
use_gitignore()
}
\value{
nothing, creates a file.
}
\description{
Use a starter .gitignore
}
\details{
Drop a starter .gitignore in the current working directory, including
ignores for drake and capsule (renv).
}
\author{
Miles McBain
}
