% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tflow.R
\name{use_rmd}
\alias{use_rmd}
\title{use_rmd}
\usage{
use_rmd(target_name)
}
\arguments{
\item{target_name}{a name for target and the generated R markdown document.}
}
\value{
the path of the file created. (invisibly)
}
\description{
Create an RMarkdown file and generate target definition code.
}
\details{
The generated document defaults to the "./doc" folder. This can be overridden
with option 'tflow.report_dir'.
}
\author{
Miles McBain
}
