% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dflow.R
\name{use_rmd_target}
\alias{use_rmd_target}
\title{use_rmd_target}
\usage{
use_rmd_target(target_file = NULL)
}
\arguments{
\item{target_file}{}
}
\value{
target text to the console.
}
\description{
Generate a target for an R markdown file
}
\details{
Due to the way RMarkdown and Knitr use relative paths to the source document
it can be messy to properly tag the input and output documents for an Rmd target. This
function will generate a multi-expression target using `drake::target()`
that uses a clean and simple way to mark these up for drake.

The contents of the `file_out()` call my need to be modified depending on
the output file extension and path configured in the Rmd and call to `render()`.
}
\author{
Miles McBain
}
