% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tflow.R
\name{use_tflow}
\alias{use_tflow}
\title{use_tflow}
\usage{
use_tflow()
}
\value{
Nothing. Modifies your workspace.
}
\description{
Setup a tflow project
}
\details{
Creates files and directories according to the tflow template.
}
