% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tflow.R
\name{rmd_target}
\alias{rmd_target}
\title{rmd_target}
\usage{
rmd_target(target_name)
}
\arguments{
\item{name}{of a target to generate rmd target for.}
}
\value{
target text to the console.
}
\description{
Generate a target for an R markdown file
}
\author{
Miles McBain
}
