% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dflow.R
\name{use_rmd}
\alias{use_rmd}
\title{use_rmd}
\usage{
use_rmd(target_file)
}
\arguments{
\item{target_file}{a filename for the generated R markdown document.}
}
\value{
the path of the file created. (invisibly)
}
\description{
Create an RMarkdown file and generate target definition code.
}
\details{
The generated document defaults to the "./doc" folder. This can be overridden
with option 'dflow.report_dir'.

Due to the way RMarkdown and Knitr use relative paths to the source document
it can be messy to properly tag the input and output documents for an Rmd target. This
function will generate a multi-expression target using `drake::target()`
that uses a clean and simple way to mark these up for drake.

The contents of the `file_out()` call my need to be modified depending on
the output file extension and path configured in the Rmd and call to `render()`.
}
\author{
Miles McBain
}
